requirements.txt
cryptography
paho-mqtt
pydantic